Get dependencies as a pre-action instead of target
Generating the xcconfig as a target meant having to run the project
twice (once to generate the config and have the build fail, and another
time to actually pull the config values and build properly). Running as
a pre-action allows the config values to be pulled on the same build.